Changed price type from Double to BigDecimal Added unit test for OrderSuccessful view model

+++ data/src/main/java/com/harutyun/data/mapper/PizzaMapper.kt
@@ -12,7 +12,7 @@ class PizzaMapper {
     fun mapToDomain(input: PizzaEntity): Pizza {
         return Pizza(
             name = input.name,
-            price = input.price,
+            price = input.price.toBigDecimal(),
         )
     }
 }
